Move Options to a separate component
