+++ nextjs/components/Channel/ChannelRow/Options/index.module.scss
@@ -1,0 +1,19 @@
+@import 'styles/colors.scss';
+
+.options {
+  display: flex;
+  padding: 0;
+}
+
+.options li {
+  padding: 0.375rem;
+  font-size: 16px;
+}
+
+.options li:hover {
+  background: rgb(249 250 251);
+}
+
+.pinned {
+  color: $color-secondary;
+}
