+++ nextjs/components/Channel/ChannelRow/index.module.scss
@@ -9,7 +9,6 @@
   border-radius: 4px;
   border: 1px solid $color-border;
   box-shadow: 0 3px 5px $color-border;
-  position: absolute;
   display: none;
   top: 0;
   right: 0;
@@ -28,8 +27,4 @@
 .container:hover .options {
   align-items: center;
   display: flex;
-}
-
-.pinned {
-  color: $color-secondary;
 }
